Add BuildInsightsPrinter to encapsulate CLI printing logic
