+++ components/fetch-build-scan-data-cmdline-tool/src/main/java/com/gradle/enterprise/cli/FetchBuildScanDataCommand.java
@@ -2,7 +2,6 @@ package com.gradle.enterprise.cli;
 
 import com.gradle.enterprise.api.client.FailedRequestException;
 import com.gradle.enterprise.api.client.GradleEnterpriseApiClient;
-import com.gradle.enterprise.model.BuildTimeMetrics;
 import com.gradle.enterprise.model.BuildScanData;
 import com.gradle.enterprise.model.CustomValueNames;
 import com.gradle.enterprise.model.NumberedBuildScan;
@@ -71,7 +70,7 @@ public class FetchBuildScanDataCommand implements Callable<Integer> {
         logFinishedFetchingAllBuildScanData();
         logFetchResults(buildScanData, customValueKeys);
 
-        printResults(buildScanData);
+        BuildInsightsPrinter.printInsights(buildScanData);
 
         return ExitCode.OK;
     }
@@ -172,46 +171,6 @@ public class FetchBuildScanDataCommand implements Callable<Integer> {
             "%s %s from custom value with name '%s' for %s build",
             found ? "Found": "Did not find", property, customValueKey, toOrdinal(runNum)
         );
-    }
-
-    private static void printResults(List<BuildScanData> buildScanData) {
-        printBuildScanDataHeader();
-        printBuildScanData(buildScanData);
-
-        if (buildScanData.size() == 2) {
-            printBuildTimeMetricsHeader();
-            printBuildTimeMetrics(buildScanData);
-        }
-    }
-
-    public static void printBuildScanDataHeader() {
-        List<String> labels = BuildScanDataFields.ordered().map(f -> f.label).collect(Collectors.toList());
-        System.out.println(String.join(",", labels));
-    }
-
-    private static void printBuildScanData(List<BuildScanData> buildScanData) {
-        buildScanData.forEach(FetchBuildScanDataCommand::printBuildScanData);
-    }
-
-    private static void printBuildScanData(BuildScanData buildScanData) {
-        List<String> values = BuildScanDataFields.ordered().map(f -> f.value.apply(buildScanData)).collect(Collectors.toList());
-        System.out.println(String.join(",", values));
-    }
-
-    private static void printBuildTimeMetricsHeader() {
-        List<String> labels = BuildTimeMetricsFields.ordered().map(f -> f.label).collect(Collectors.toList());
-        System.out.println(String.join(",", labels));
-    }
-
-    private static void printBuildTimeMetrics(List<BuildScanData> buildScanData) {
-        final BuildTimeMetrics buildTimeData = BuildTimeMetrics.from(buildScanData.get(0), buildScanData.get(1));
-        List<String> values;
-        if (buildTimeData == null) {
-            values = BuildTimeMetricsFields.ordered().map(f -> "").collect(Collectors.toList());
-        } else {
-            values = BuildTimeMetricsFields.ordered().map(f -> f.value.apply(buildTimeData)).collect(Collectors.toList());
-        }
-        System.out.println(String.join(",", values));
     }
 
     private static String toOrdinal(int i) {
